#  Store `mmMaxTime` in same field as `seriesShard`