We don't use `seriesShard` during DB initialization, so we can use the same 8 bytes to store `mmMaxTime`, and save those during the rest of the lifetime of the database.

Related PRs:
- https://github.com/prometheus/prometheus/pull/14472
- https://github.com/prometheus/prometheus/pull/14473
- https://github.com/prometheus/prometheus/pull/9160 (this is where `mmMaxTime` was introduced).

I also added a benchmark that runs on a _real_ WAL, and tried with a 5 GB WAL like this:
```
$ du -h
0	./wbl
376M	./wal/checkpoint.00078035
5.4G	./wal
6.3G	./chunks_head
12G	.
```

The result shows no difference in CPU time, as expected (ran on an M3 Pro Apple Silicon) 
```
goos: darwin
goarch: arm64
pkg: github.com/prometheus/prometheus/tsdb
               │    main    │           inshard            │
               │   sec/op   │   sec/op     vs base         │
LoadRealWLs-12   99.82 ± 6%   100.43 ± 5%  ~ (p=1.000 n=6)

               │     main     │            inshard            │
               │     B/op     │     B/op      vs base         │
LoadRealWLs-12   34.99Gi ± 0%   35.03Gi ± 1%  ~ (p=0.394 n=6)

               │    main     │           inshard            │
               │  allocs/op  │  allocs/op   vs base         │
LoadRealWLs-12   382.8M ± 0%   382.8M ± 0%  ~ (p=0.699 n=6)
```

This reduces the `unsafe.Sizeof(memSeries{})` from 168 to 160 with `stringlabels` (176->168 without tags and with `dedupelabels`), which would move it to a [160-byte class](https://go.dev/src/runtime/sizeclasses.go) with `stringlabels`, saving 16 bytes per series.